requirements.txt
aiohttp
python-axolotl
cryptography
